+++ selling-used-items-app-backend/selling-used-items-app-backend/Controllers/CommentController.cs
@@ -1,6 +1,8 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using selling_used_items_app_backend.Model;
 using selling_used_items_app_backend.Service;
+using selling_used_items_app_backend.Validator.CommentValidator;
+using System.ComponentModel.DataAnnotations;
 
 namespace selling_used_items_app_backend.Controllers
 {
@@ -9,10 +11,14 @@ namespace selling_used_items_app_backend.Controllers
     public class CommentController : ControllerBase
     {
         private readonly ICommentService _commentService;
+        private readonly CommentCreateValidator _commentCreateValidator;
+        private readonly CommentDeleteValidator _commentDeleteValidator;
 
-        public CommentController(ICommentService commentService)
+        public CommentController(ICommentService commentService, CommentDeleteValidator commentDeleteValidator, CommentCreateValidator commentCreateValidator)
         {
             _commentService = commentService ?? throw new ArgumentNullException(nameof(commentService));
+            _commentCreateValidator = commentCreateValidator;
+            _commentDeleteValidator = commentDeleteValidator;
         }
 
         [HttpGet]
@@ -36,25 +42,23 @@ namespace selling_used_items_app_backend.Controllers
         [HttpPost]
         public IActionResult CreateComment(Comment comment)
         {
+            var validationResult = _commentCreateValidator.ValidateComment(comment);
+            if (validationResult != ValidationResult.Success)
+            {
+                return BadRequest(validationResult.ErrorMessage);
+            }
             _commentService.Create(comment);
             return CreatedAtAction(nameof(GetComment), new { id = comment.id }, comment);
-        }
-
-        [HttpPut("{id}")]
-        public IActionResult UpdateComment(int id, Comment comment)
-        {
-            if (id != comment.id)
-            {
-                return BadRequest();
-            }
-
-            _commentService.Update(comment);
-            return NoContent();
         }
 
         [HttpDelete("{id}")]
         public IActionResult DeleteComment(int id)
         {
+            var validationResult = _commentDeleteValidator.ValidateComment(id);
+            if (validationResult != ValidationResult.Success)
+            {
+                return NotFound(validationResult.ErrorMessage);
+            }
             _commentService.Delete(id);
             return NoContent();
         }
