+++ selling-used-items-app-backend/selling-used-items-app-backend/Controllers/UserController.cs
@@ -1,6 +1,8 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using selling_used_items_app_backend.Model;
 using selling_used_items_app_backend.Service;
+using selling_used_items_app_backend.Validator.UserValidator;
+using System.ComponentModel.DataAnnotations;
 
 namespace selling_used_items_app_backend.Controllers
 {
@@ -9,10 +11,17 @@ namespace selling_used_items_app_backend.Controllers
     public class UserController : ControllerBase
     {
         private readonly IUserService _userService;
+        private readonly UserCreateValidator _createUserValidator;
+        private readonly UserUpdateValidator _updateUserValidator;
+        private readonly UserDeleteValidator _deleteUserValidator;
 
-        public UserController(IUserService userService)
+
+        public UserController(IUserService userService, UserCreateValidator createUserValidator, UserUpdateValidator updateUserValidator, UserDeleteValidator deleteUserValidator)
         {
-            _userService = userService ?? throw new ArgumentNullException(nameof(userService));
+            _userService = userService;
+            _createUserValidator = createUserValidator;
+            _updateUserValidator = updateUserValidator;
+            _deleteUserValidator = deleteUserValidator;
         }
 
         [HttpGet]
@@ -36,6 +45,11 @@ namespace selling_used_items_app_backend.Controllers
         [HttpPost]
         public IActionResult CreateUser(User user)
         {
+            var validationResult = _createUserValidator.ValidateUser(user);
+            if (validationResult != ValidationResult.Success)
+            {
+                return BadRequest(validationResult.ErrorMessage);
+            }
             _userService.Create(user);
             return CreatedAtAction(nameof(Get), new { id = user.id }, user);
         }
@@ -45,9 +59,13 @@ namespace selling_used_items_app_backend.Controllers
         {
             if (id != user.id)
             {
-                return BadRequest();
+                return BadRequest("ID in the request path does not match the ID in the request body.");
             }
-
+            var validationResult = _updateUserValidator.ValidateUser(user);
+            if (validationResult != ValidationResult.Success)
+            {
+                return BadRequest(validationResult.ErrorMessage);
+            }
             _userService.Update(user);
             return NoContent();
         }
@@ -55,8 +73,14 @@ namespace selling_used_items_app_backend.Controllers
         [HttpDelete("{id}")]
         public IActionResult DeleteUser(int id)
         {
+            var validationResult = _deleteUserValidator.ValidateUser(id);
+            if (validationResult != ValidationResult.Success)
+            {
+                return NotFound(validationResult.ErrorMessage);
+            }
             _userService.Delete(id);
             return NoContent();
         }
+
     }
 }
